Implementing interfaces.
A simple example to help understand how interfaces can be implemented.
Here, the drive method is implemented through the Car interface.
Other sample interfaces such as Loggable, Origins and Specs were also
created, which can be used similar to Car. The abstract methods defined
inside the interface will be used in the classes which implement the
interface.
Example: drive is an abstract method defined inside the Car interface.
This method can now be used in Audi, Ferrari and Mercedes classes to
implement thier specific drive functionality.