Finalize method of class Object
Added a simple example to understand the finalize method of the Object
class. Since the programmer does not have explicit control over the
finalize method, it is recommended not to use it. Instead, to close any
open resources, it is recommended to have a custom method and call it
inside a try finally to release resources.

The finalize is called by the JVM asynchronously during garbage
collection.