Encapsulation sample code.
To demonstrate the concept of encapsulation in java, private members
were created and these members were fetched/set using setter and getter
methods.
When the members are accessed directly (statically) using the object of
the class, we get a variable not visible compilation error.
The basic idea of encapsulation is to restrict direct access of these
members from other classes.