Data serialization and deserialization
The state of the object can be saved even after run time, by serializing
the data (object). The data (object) when serialized, gets stored in a
.ser file, which can later be used to retrieve original values
(deserializing).
Serializing - Convert object to byte stream.
Deserializing - Convert byte stream to object.
Every byte stream of a class' object will have a SerialVersionUID. The
same is used while deserializing, the reason for why the
SerialVersionUID has to remain the same while serializing and
deserializing. This UID can also be defined implicitly inside the class
whose data will be serialized.