Enums and recursion
To demonstrate the use of enums, the planets in our solar system was
considered. These values were also indexed, to show how the associated
values in enums can be retrieved/used.
To demonstrate the concept of recursion, the idea of staying hydrated
was used. The program asks for number of reminders to give and the
quantity of water to drink and prints out the status.