Standard exceptions and custom exceptions
To show how a standard pre-defined java exception can be caught,
division of a number by zero was executed which results in a
ArithmeticException.
To show how a custom exception can be written, an example of registering
a team with a minimum of 15 players and a maximum of 25 players was
written. If the number was outside of this range, it results in a
PlayerLimitException which is user-defined.